Refactor: changed folder structure to different names and additional index.js
