+++ Client/src/components/Filter/SearchFilter.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Button from '../Button';
+import { Button } from '../Button/';
 
 const Searchfilter = ({ onSearchSubmit }) => {
 	const [searchInput, setSearchInput] = useState('');
