+++ Client/src/components/Genres/index.js
@@ -1,14 +1,3 @@
-import React from 'react';
-import './Genres.css'
+import Genres from "./Genres";
 
-const Genres = ({ genres }) => {
-	return (
-		<>
-			{genres.map(
-				genre => <li key={genre} className="rounded-border-0 movie-genre">{genre}</li>
-			)}
-		</>
-	);
-}
-
-export default Genres;
+export { Genres };
