+++ Client/src/components/Image/index.js
@@ -1,19 +1,3 @@
-import React from 'react';
-import './Image.css'
-import fallbackImg from '../../assets/fallback.png'
+import Image from './Image';
 
-const Index = ({ src, alt, className }) => {
-	return (
-		<img 
-			src={src}
-			loading="lazy" 
-			alt={alt} 
-			className={className}
-			onError={
-				(e) => {e.target.src = fallbackImg}
-			} 
-		/>
-	);
-}
-
-export default Index;
+export { Image };
