+++ Client/src/components/MoviesList/index.js
@@ -1,15 +1,3 @@
-import React from 'react';
-import Card from '../Card'
-import './Movies.css'
+import MoviesList from "./MoviesList";
 
-const Index = ({movies}) => {
-	return (
-		<div className='movies d-grid justify-center gap-1'>
-			{movies.map( movie => 
-				<Card key={movie.id} movie={movie} /> 
-			)}
-		</div>
-	);
-}
-
-export default Index;
+export { MoviesList };
